41 reducer
